Refactor: defaults, safety checks, and cleanups
Apply small refactors and safety/clarity fixes across crates: derive Default for several enums/structs and remove manual impls; tighten bounds checks using ranges/.contains and clamp; simplify conditionals (matches, is_some_and, unwrap_or_default); replace manual loops with iter/enumerate and std helpers (from_fn, size_of_val, div_ceil); minor API rename (Difficulty::from_str -> parse) and added convenience methods (SoundRegistry::is_empty); fix path/dir handling and integer/float arithmetic formatting; move/adjust tests and tidy UI/renderer code. Improves correctness, readability and eliminates redundant code.

+++ crates/game/src/chat.rs
@@ -580,9 +580,9 @@ impl ChatState {
                         // count rather than the bare extent — otherwise a
                         // user asking for "32 blocks wide" would silently get
                         // 33.
-                        let cell_x = (x1 - x2).unsigned_abs() as u32 + 1;
-                        let cell_y = (y1 - y2).unsigned_abs() as u32 + 1;
-                        let cell_z = (z1 - z2).unsigned_abs() as u32 + 1;
+                        let cell_x = (x1 - x2).unsigned_abs() + 1;
+                        let cell_y = (y1 - y2).unsigned_abs() + 1;
+                        let cell_z = (z1 - z2).unsigned_abs() + 1;
                         if cell_x > 32 || cell_y > 32 || cell_z > 32 {
                             return CommandResult::Unknown(format!(
                                 "/copy selection too large: \
